Extract mock location helper function from navigation tests.
